Fix static analysis issues
:Release Notes:
Fix static analysis issues

:Detailed Notes:
Fixed below static issues:
CID-9168477 CERT-CPP Characters and Strings
CID-9168476 CERT-C Characters and Strings
CID-9168469 CERT-C Characters and Strings
CID-9168465 CERT-CPP Characters and Strings
CID-9168457 CERT-CPP Characters and Strings
CID-9168445 CERT-CPP Characters and Strings
CID-9168444 CERT-CPP Characters and Strings
CID-9168436 CERT-C Characters and Strings
CID-9168432 CERT-CPP Characters and Strings
CID-9168426 CERT-CPP Characters and Strings
CID-9156754 CERT-C Expression
CID-9156746 CERT-C Expression
CID-9156739 CERT-C Expression
CID-9156729 Dereference null return value
CID-9156727 Dereference null return value
CID-9156726 CERT-C Expression
CID-9156723 CERT-C Characters and Strings
CID-9156721 Dereference null return value
CID-9156706 Dereference null return value
CID-9044517 CERT-CPP Characters and Strings
CID-9044478 CERT-CPP Characters and Strings
CID-9044471 CERT-CPP Characters and Strings

:Testing Performed:
Verified in local build

:QA Notes:

:Issues Addressed:
[WRP-8420] Fix static issues of notificationmgr for OSE

+++ src/PincodeValidator.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2016-2018 LG Electronics, Inc.
+// Copyright (c) 2016-2023 LG Electronics, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -29,8 +29,8 @@ static std::string bin2hex(const unsigned char *bin, size_t len)
     for(size_t i = 0; i < len; ++i)
     {
         unsigned char c = (unsigned char)bin[i];
-        res += hex[c >> 4];
-        res += hex[c & 0xf];
+        res += hex[(unsigned int)(c >> 4)];
+        res += hex[(unsigned int)(c & 0xf)];
     }
 
     return res;
